consumer/requirements.txt
kafka-python-ng
psycopg2-binary
python-dateutil==2.9.0.post0
joblib==1.4.2
scikit-learn==1.5.0
pandas
mlflow
python-dotenv
requests
pydantic-settings==2.6.1